+++ projects/triangular-engine/src/lib/engine/components/engine-ui/engine-ui.component.ts
@@ -1,21 +1,20 @@
-import { Component, TemplateRef, input } from '@angular/core';
+import { Component, input } from '@angular/core';
+import { CommonModule } from '@angular/common';
 
 import { EngineService } from '../../services';
-
-import { CommonModule } from '@angular/common';
 import { EngineStatsComponent } from './engine-stats/engine-stats.component';
 import { IUserInterfaceOptions } from '../../models';
 import { SceneTreeComponent } from '../../features/scene-tree/scene-tree.component';
+import { EnginePortalOutletComponent } from './engine-portal-outlet.component';
 
 @Component({
-    selector: 'engine-ui',
-    imports: [CommonModule, EngineStatsComponent, SceneTreeComponent],
-    templateUrl: './engine-ui.component.html',
-    styleUrl: './engine-ui.component.scss'
+  selector: 'engine-ui',
+  imports: [CommonModule, EngineStatsComponent, SceneTreeComponent, EnginePortalOutletComponent],
+  templateUrl: './engine-ui.component.html',
+  styleUrl: './engine-ui.component.scss'
 })
 export class EngineUiComponent {
   readonly userInterface = input<IUserInterfaceOptions>({});
-  //
 
   constructor(readonly engineService: EngineService) {}
 }